requirements.txt
asgiref==3.6.0
channels-redis==4.1.0
Django==4.2.1
django-environ==0.10.0
psycopg2==2.9.6
sqlparse==0.4.4
